Moved tracking and prefs to core
