+++ WooCommerce/src/main/kotlin/com/woocommerce/android/di/SiteComponent.kt
@@ -2,14 +2,9 @@ package com.woocommerce.android.di
 
 import androidx.datastore.core.DataStore
 import com.woocommerce.android.ui.mystore.data.DashboardDataModel
-import dagger.BindsInstance
-import dagger.hilt.DefineComponent
 import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineScope
-import org.wordpress.android.fluxc.model.SiteModel
-import javax.inject.Qualifier
 import javax.inject.Scope
 import kotlin.annotation.AnnotationRetention.RUNTIME
 
@@ -17,17 +12,6 @@ import kotlin.annotation.AnnotationRetention.RUNTIME
 @MustBeDocumented
 @Retention(value = RUNTIME)
 annotation class SiteScope
-
-@SiteScope
-@DefineComponent(parent = SingletonComponent::class)
-interface SiteComponent {
-    @DefineComponent.Builder
-    interface Builder {
-        fun setSite(@BindsInstance site: SiteModel): Builder
-        fun setCoroutineScope(@BindsInstance @SiteCoroutineScope scope: CoroutineScope): Builder
-        fun build(): SiteComponent
-    }
-}
 
 @InstallIn(SiteComponent::class)
 @EntryPoint
@@ -37,8 +21,3 @@ interface SiteComponentEntryPoint {
     @SiteCoroutineScope
     fun siteCoroutineScope(): CoroutineScope
 }
-
-@Qualifier
-@MustBeDocumented
-@Retention(RUNTIME)
-annotation class SiteCoroutineScope
