+++ WooCommerce/src/main/res/layout/fragment_developer_options.xml
@@ -21,6 +21,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="POS"
+        tools:ignore="HardcodedText"
         app:layout_constraintTop_toBottomOf="@id/developerOptionsRv"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
@@ -32,6 +33,7 @@
         android:text="POS Internal"
         app:layout_constraintTop_toBottomOf="@id/posButton"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
+        app:layout_constraintEnd_toEndOf="parent"
+        tools:ignore="HardcodedText" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
