+++ libs/core/src/main/java/com/woocommerce/android/analytics/AnalyticsEvent.kt
