+++ libs/core/src/main/java/com/woocommerce/android/extensions/PackageManagerExt.kt
