+++ libs/core/src/main/java/com/woocommerce/android/extensions/StringExt.kt
