+++ libs/core/src/main/java/com/woocommerce/android/tools/SelectedSite.kt
