+++ libs/core/src/main/java/com/woocommerce/android/tools/SiteConnectionType.kt
@@ -1,6 +1,6 @@
 package com.woocommerce.android.tools
 
-import com.woocommerce.android.BuildConfig
+import com.woocommerce.android.core.BuildConfig
 import com.woocommerce.android.util.WooLog
 import org.wordpress.android.fluxc.model.SiteModel
