+++ libs/core/src/main/java/com/woocommerce/android/util/DeviceInfo.kt
