+++ libs/core/src/main/java/com/woocommerce/android/util/PreferenceUtils.kt
