+++ libs/core/src/main/java/com/woocommerce/android/util/RollingLogEntries.kt
