+++ libs/pos/build.gradle
@@ -85,7 +85,6 @@ dependencies {
     implementation 'androidx.compose.material3:material3'
     implementation 'androidx.navigation:navigation-compose:2.7.7'
 
-    implementation "com.automattic:Automattic-Tracks-Android:$automatticTracksVersion"
     implementation "com.automattic.tracks:crashlogging:$automatticTracksVersion"
 
     implementation "com.google.dagger:hilt-android:$gradle.ext.daggerVersion"
